requirements.txt
asgiref==3.6.0
attrs==22.2.0
colorama==0.4.6
Django==4.1.5
django-filter==22.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
exceptiongroup==1.1.0
iniconfig==2.0.0
Markdown==3.4.1
packaging==23.0
Pillow==9.4.0
pluggy==1.0.0
psycopg2==2.9.5
PyJWT==2.6.0
pytest==7.2.0
pytest-django==4.5.2
python-decouple==3.7
pytz==2022.7
sqlparse==0.4.3
tomli==2.0.1
tzdata==2022.7
